Channel update page routing & Fixed conflicts (#33)
* Initialze Add posts & Add coments functions

* Comment action.ts for test

* Clean Conflicts

* Implement Edit & Delete functions

* save

* fixed profile cover image and text color

* update password complete

* completed styling

* merge with chenni

* Takara/fix styling (#30)

* removed dark mode:

* fixed profile page styling for modal

* updated styling and fixed modal
:

---------


* Initialze Add posts & Add coments functions

* Comment action.ts for test

* Clean Conflicts

* Implement Edit & Delete functions

* Rename page routing

* Cleaned up dir

* Merged with Takara & cleaned up dir

* Cleaned up current channel page

* Update AllGroups.tsx

* Update Groups.tsx

* Update Members.tsx

* Update actions.ts

* Update middleware.tsx

* Update page.tsx

* Update page.tsx

* Update page.tsx

* Update fetchCurChannel.ts

* Delete src/app/dashboard/channels/currentChannel/[id]/[userId]/page.tsx

* Delete src/app/dashboard/channels/members_detail/[id]/[userId]/page.tsx

---------
