+++ src/components/channelContents/AllGroups.tsx
@@ -12,24 +12,24 @@ type User = {
     last_name: string | null;
     organization: string | null;
     title: string | null;
-    phone: string | null;
+    phone: string | null; 
     description: string | null;
     password: string | null;
     personal_email: string | null;
     graduation_year: string | null;
     work_email: string | null;
     createdAt: Date;
-};
-
-type Comment = {
+  };
+  
+  type Comment = {
     id: number;
     desc: string;
     userId: string;
     postId: number;
     user: User;
-};
-
-type Post = {
+  };
+  
+  type Post = {
     id: number;
     desc: string;
     img: string | null;
@@ -40,18 +40,18 @@ type Post = {
     channelId: number;
     user: User;
     comments: Comment[];
-};
-
-type Channel = {
+  };
+  
+  type Channel = {
     id: number;
     channel_name: string;
     channel_image: string | null;
     channel_description: string | null;
     users: {
-        user: User;
+      user: User;
     }[];
     posts: Post[];
-};
+  };
 
 interface AllGroupsProps {
     channel: Channel;
@@ -75,47 +75,47 @@ const AllGroups: React.FC<AllGroupsProps> = ({ channel, currentUser }) => {
 
     useEffect(() => {
         const fetchChannels = async () => {
-            try {
-                console.log("Fetching channels for user:", userId);
-
-                const response = await fetch(
-                    `/api/channel/fetchChannels?userId=${userId}`
-                );
-                if (!response.ok) {
-                    throw new Error("Failed to fetch channels");
-                }
-
-                const data = await response.json();
-
-                setUsername(data.username);
-                setJoinedChannels(data.joinedChannels);
-                setNotJoinedChannels(data.notJoinedChannels);
-
-                console.log("User Channels:", data.joinedChannels);
-                console.log("Not Joined Channels:", data.notJoinedChannels);
-            } catch (error) {
-                console.error("Error fetching channels:", error);
+          try {
+            console.log("Fetching channels for user:", userId);
+    
+            const response = await fetch(
+                `/api/channel/fetchChannels?userId=${userId}`
+            );
+            if (!response.ok) {
+              throw new Error('Failed to fetch channels');
             }
+    
+            const data = await response.json();
+    
+            setUsername(data.username);
+            setJoinedChannels(data.joinedChannels);
+            setNotJoinedChannels(data.notJoinedChannels);
+    
+            console.log("User Channels:", data.joinedChannels);
+            console.log("Not Joined Channels:", data.notJoinedChannels);
+          } catch (error) {
+            console.error("Error fetching channels:", error);
+          }
         };
-
+    
         fetchChannels();
-    }, [userId]);
+      }, [userId]);
 
     const combinedChannels = joinedChannels.concat(notJoinedChannels);
 
     return (
         <div className="flex flex-col gap-6">
-            <div className="p-4 bg-white rounded-lg shadow-md text-lg flex flex-col gap-4">
+            <div className='p-4 bg-white rounded-lg shadow-md text-lg flex flex-col gap-4'>   
                 <div className="">
-                    <Link
-                        href={`/dashboard/channels/currentChannel/${channel.id}/${userId}`}
+                    <Link 
+                        href={`/dashboard/channels/currentChannel/${channel.channel_name}/${username}`}
                     >
                         <button>
-                            <Image
-                                src="/backArrow.png"
-                                alt="back"
-                                width={20}
-                                height={20}
+                            <Image 
+                                src="/backArrow.png" 
+                                alt="back" 
+                                width={20} 
+                                height={20} 
                             />
                         </button>
                     </Link>
@@ -123,38 +123,37 @@ const AllGroups: React.FC<AllGroupsProps> = ({ channel, currentUser }) => {
                 </div>
                 <div className="flex justify-between items-center font-medium">
                     <span className="text-gray-500">All Channels</span>
-                    <div className="flex p-2 bg-slate-100 items-center rounded-xl">
-                        <input
-                            type="text"
-                            placeholder="search..."
+                    <div className='flex p-2 bg-slate-100 items-center rounded-xl'>
+                        <input 
+                            type="text" 
+                            placeholder="search..." 
                             className="bg-transparent outline-none text-sm"
                         />
-                        <Image
-                            src="/search.png"
-                            alt=""
-                            width={14}
+                        <Image 
+                            src="/search.png" 
+                            alt="" width={14} 
                             height={14}
                         />
                     </div>
-                </div>
+                </div>                  
             </div>
-            <div className="p-4 bg-white rounded-lg shadow-md text-lg flex flex-col gap-4">
-                <ul className="px-2 text-lg flex flex-col gap-4">
+            <div className='p-4 bg-white rounded-lg shadow-md text-lg flex flex-col gap-4'>
+                <ul className='px-2 text-lg flex flex-col gap-4'>
                     {joinedChannels.map((channel) => (
                         <li key={channel.id}>
                             <div className="flex justify-between">
-                                <Link
-                                    href={`/dashboard/channels/currentChannel/${channel.id}/${userId}`}
-                                    className="flex items-center gap-4 p-2 rounded-lg hover:bg-slate-100 w-full"
+                                <Link 
+                                    href={`/dashboard/channels/currentChannel/${channel.channel_name}/${username}`}
+                                    className="flex items-center gap-4 p-2 rounded-lg hover:bg-slate-100"
                                 >
-                                    <img
-                                        src={channel.channel_image || ""}
-                                        alt={channel.channel_name}
+                                    <img 
+                                        src={channel.channel_image || ''} 
+                                        alt={channel.channel_name} 
                                         width={20}
                                         height={20}
                                         className="w-10 h-10 rounded-full"
                                     />
-                                    <span className="font-medium">
+                                    <span className="font-semibold font-medium">
                                         {channel.channel_name}
                                     </span>
                                 </Link>
@@ -164,16 +163,16 @@ const AllGroups: React.FC<AllGroupsProps> = ({ channel, currentUser }) => {
                                     </button>
                                 </Link>
                             </div>
-                            <hr className="border-t-1 border-gray-50 w-36 self-center" />
-                        </li>
-                    ))}
+                            <hr className="border-t-1 border-gray-50 w-36 self-center"/>
+                        </li>   
+                    ))}            
                 </ul>
                 <ul className="px-2 text-lg flex flex-col gap-4">
                     {notJoinedChannels.map((channel) => (
                         <li key={channel.id}>
                             <div className="flex justify-between">
                                 <Link
-                                    href={`/dashboard/channels/currentChannel/${channel.id}/${userId}`}
+                                    href={`/dashboard/channels/currentChannel/${channel.channel_name}/${username}`}
                                     className="flex items-center gap-4 p-2 rounded-lg hover:bg-slate-100"
                                 >
                                     <img
@@ -199,7 +198,9 @@ const AllGroups: React.FC<AllGroupsProps> = ({ channel, currentUser }) => {
                 </ul>
             </div>
         </div>
-    );
-};
+    )
+}
+
+export default AllGroups;
 
 export default AllGroups;
