+++ src/lib/actions.ts
@@ -527,4 +527,3 @@ export const addComment = async (desc: string, postId: number) => {
     throw new Error("Something went wrong!");
   }
 };
-
