+++ src/pages/api/channel/fetchUser.ts
@@ -3,15 +3,17 @@ import { NextApiRequest, NextApiResponse } from 'next';
 import prisma from '@/lib/client';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
-  const { userId } = req.query;
+  const { userName } = req.query;
 
-  if (typeof userId !== 'string') {
-    return res.status(400).json({ error: 'Invalid userId' });
+  console.log(userName)
+
+  if (typeof userName !== 'string') {
+    return res.status(400).json({ error: 'Invalid username' });
   }
 
   try {
     const user = await prisma.user.findUnique({
-      where: { id: userId },
+      where: { username: userName },
     });
 
     if (!user) {
